Fix #15: get threshold exceeding users blocking stable
We should update threshold on the previous data (users_before), not
current one (users_after).

There is also a subtle race possible: if we take 2 adjacent periods from
the database, then we likely see the same user in both of them, so model
validation always failed under heavy load. Probably this is because of
delays on many layers, from tfw_logger and clickhouse to Ptyhon
processing on it's own.

NOTICE: I fixed the current logic, but basically the current
TRAINING_MODE="off" is more or less the same as "real" with only
exception that we adjust parameters on live sliding window instead of
single historical observation. I'd expect from this mode to rely on
explicitly defined thresholds.